[Telemetry] Add Expectation module that enables disabling benchmarks/stories
This work is for better disabling of telemetry tests. This first
CL lays the ground works by creating the story_expectation object
and by extending the benchmark class api to include a way to
get story expectations. This implements the default behavior of creating
an empty expectation set if none is specified.

go/StoryExpectationsDD

BUG=chromium:711065

Review-Url: https://codereview.chromium.org/2843403005