[Telemetry] Add ability to check if device has AOSP build to android platform backend.
This will be required to add a TestExpectation TestCondition for determining if we have
a webview build. The TestConditions are only passed the platform. This does not
contain information about the browser type. Webview bots have to run using an AOSP build,
and no other telemetry bots run aosp builds.

BUG=chromium:713222

Review-Url: https://codereview.chromium.org/2904163002