[pinpoint] Rename isolated to isolate.
"Isolated" is pretty awkward-sounding as far as English grammar goes,
so its usage is typo-prone.

In addition, "isolated" refers to the JSON file containing the isolate
information. Pinpoint is not aware of these JSON files, and the Isolate Server
doesn't store data in that format anyway. It's better to just use "isolate" to
refer to the "Isolate Server".

"[The Isolate Server] is ill named, it doesn't know about the .isolate or
.isolated file formats."
https://github.com/luci/luci-py/blob/master/appengine/isolate/doc/Design.md
Review-Url: https://codereview.chromium.org/2903383004