Cpu Time Tree Data Reporter
The added reportCpuTimeTreeValuesToHistogramSet functions traverses a
cpu time multidimensional view, and reports data from it as single
valued histograms to a provided histogramSet.

The module defines a private class CpuTimeTreeDataReporter, which
encapsulates the necessary state (e.g. keeping track of which nodes has
been visited) for the tree traversal.

This CL was spun off from https://codereview.chromium.org/2900073002/,
and the code added will be used for the new cpu time metric.

BUG=catapult:#3325

Review-Url: https://codereview.chromium.org/2906663003