Revert of Set BOTO_CONFIG correctly in cloud_storage.py (patchset #1 id:1 of https://codereview.chromium.org/2923123006/ )
Reason for revert:
It turns out this did not fix the issue on the buildbot. The .boto file might be on the buildbot at this path, but the tests that are failing are running inside a VM, and don't have access to it in the first place.

Another approach is needed to fix the linked bug, and we will likely put a .boto file in the home directory that is being set when running gsutil.

Original issue's description:
> Set BOTO_CONFIG correctly in cloud_storage.py
>
> When running the telemetry perf tests on a VM host as root, the .boto was not
> being found.
>
> This patch simply sets the BOTO_CONFIG environment variable to point at
> the correct location.
>
> Bug: chromium:712349
> Review-Url: https://codereview.chromium.org/2923123006
> Committed: https://chromium.googlesource.com/external/github.com/catapult-project/catapult/+/fbed86bbbe2593d151328e3304eddeecd0c7adfb

TBR=nednguyen@google.com
# Not skipping CQ checks because original CL landed more than 1 days ago.

Review-Url: https://codereview.chromium.org/2931173002