Fix issues that caused events to render inaccurately at high zoom levels.
I’ve attached a file to bug #625 that demonstrates the issue for the
various types of event affected.

I speculate that the issues occur because, somewhere in the code path of
CanvasRenderingContext2D.transform(), the elements of the transform
matrix are converted from double precision to single precision; this
would be consistent with the zoom level at which the issues start
occurring and the number of significant digits required at that zoom
level. However, I haven’t been able to find a place in the code where
such a conversion occurs.

In any case, the solution is to avoid applying the transform to the
canvas and, instead, perform the transformation explicitly using
dt.xWorldToView().

I have deleted TimelineDisplayTransform.applyTransformToCanvas() because
any use of it in the future would likely cause similar issues.

BUG=catapult:#625

Review-Url: https://codereview.chromium.org/2917083002