Implement Alert grouping on the dashboard frontend.
Currently, Alerts are grouped into AlertGroups in the backend as they are created.
This approach has a number of drawbacks described in
https://github.com/catapult-project/catapult/issues/3651

This CL takes a first step towards removing AlertGroup from the backend by adding
an opt-in switch to alerts-table to ignore the backend AlertGroups and re-group
the alerts using a similar algorithm in the frontend.

This significantly reduces the number of alert groups in the table due to known
inconsistency bugs in the backend grouping algorithm as described in 3651.

Grouping 500 alerts takes 22ms on my desktop.

This work turned up a related bug: this table is rendered multiple times:
https://github.com/catapult-project/catapult/issues/3659

Demo: https://dev-benjhayden-4df5c6c4-dot-chromeperf.appspot.com/alerts?3651=y

BUG=catapult:#3652

Review-Url: https://codereview.chromium.org/2945473002