Remove the concept of extra columns from alerts-table.
This is no longer needed now that data stoppage alerts have been
removed. Unfortunately, a side affect of this is that bug columns no
longer show on the releasing table, but we should find a more
maintainable way to add that feature back in when we work actively
on it.

BUG=catapult:#3692

Review-Url: https://codereview.chromium.org/2968013002