Dashboard - Add pinpoint_service
Continuing with dashboard integration, need to be able to route requests to Pinpoint so separating that out to a service. This will be used initially by the start_pinpoint_job handler when submitting job requests. Eventually this can be expanded to get info on job status, etc. when the dashboard tracks and does bug updates itself.

BUG=catapult:#3333

Review-Url: https://codereview.chromium.org/2976923003