Fix object-instance-track heading.
Currently, object-instance-track's heading displays the typeName of the first objectInstance.
http://i.imgur.com/Eb9z6kal.png
However, it is possible for subsequent instances in the track to have a
different typeName because object instances are grouped by baseTypeName.

This CL makes object-instance-track's heading display the baseTypeName, which is
the same for all instances in the track.
http://i.imgur.com/bapLM8kl.png
Review-Url: https://codereview.chromium.org/2987173002