Allow the user to copy tooltip links to the clipboard.
This was motivated by a request to copy the trace link, but I think if
an external link is worth showing it's worth making it easy to copy.

Also update the title of the Trace link, since users are having trouble
differentiating it from the "Trace" button.

A clearer name for the trace button, as well as converting it to use the
new --extra-chrome-categories and --extra-atrace-categories arguments
are slated for a future CL.

BUG=catapult:#3721

Review-Url: https://codereview.chromium.org/2993823002