Plumb SparseDiagnostics by Name
This CL adds the creation of Sparse Diagnostics in add_histograms by using
a set of reserved names.

It also adds some of the logic so that when shared diagnostics of histograms
are imported, their name is checked against the reserved set, instead of
only using their type.

This addresses step 5 of the list of tasks to be done in bug 3507.

BUG=catapult:#3718, catapult:#3507

Review-Url: https://codereview.chromium.org/2989143002
Committed: https://chromium.googlesource.com/external/github.com/catapult-project/catapult/+/9e5700f4e956522b4566049f0160b0dc1998c232
Review-Url: https://codereview.chromium.org/2989143002