[Telemetry] Reuse shared state between story runs
This CL removes the benchmark methods:

- ShouldTearDownStateAfterEachStoryRun
- ShouldTearDownStateAfterEachStorySetRun

allowing the shared state to be reused across story runs.

It also provides a new default implementation for the
shared state to always close the browser after each story
(except for ChromeOs).

This is not expected to change performance metrics
(although it might), and should reduce total benchmark
running time.

BUG=chromium:748566,chromium:746251

Review-Url: https://codereview.chromium.org/2998773002