Telemetry: fix run_benchmark crash on CrOS
run_benchmark against a remote CrOS target crashes because it tries to
create two connections to obtain trace data - one against the login
screen, one against the instance of Chrome running the benchmark. This
causes the second connection to fail with an error saying "Tracing
already started".

Disable the first trace connection against the login screen, as it is
not needed.

BUG=catapult:#756342

Test: Use run_benchmark to verify that this patch solves the issue;
see crbug.com/756342 for details.
Review-Url: https://codereview.chromium.org/3002793002