Fix for chromeos enterprise tests.
ChromeOS enterprise tests restart the browser. We no longer have a fixed
devtools port, so each restart requires us to reconnect to devtools,
as in this CL: https://codereview.chromium.org/2993363002/

In the past, the devtools client continued to work, but it no longer
does, and the test now needs to invoke WaitForBrowserToComeUp to
reconnect to dev tools.

BUG=chromium:755621
TEST=enterprise_RemoraRequisition

Review-Url: https://codereview.chromium.org/3005543002