Add function to calculate First Interactive.
We split the old TTI definition into First Interactive and Consistently
Interactive. This CL adds a function to calculate First Interactive.

The definition was introduced in https://goo.gl/veegmW. The lighthouse folks
found it's cleaner to implement if we introduce the concept of TaskClusters
instead of defining in terms of Lonely Tasks. See https://goo.gl/1a1XwZ for the
reconceptualized definition.

BUG=catapult:#3635

Review-Url: https://codereview.chromium.org/3012453002