[pinpoint] Adjust chart bounds and pulse size.
When a data series has outliers, showing the full data can lead the graph to be
zoomed very far out. Instead, we trim 10% off the bottom and top, and calculate
the bounds with the remaining 80%. To prevent this from cutting off the data
when there aren't outliers, we then zoom out by 20%.

Also, double the pulse size. It was a little hard to notice if you weren't
looking for it.

Review-Url: https://chromiumcodereview.appspot.com/3009943002