[pinpoint] Increase significance level.
With a separate Swarming task for each test run, we may reduce the inter-task
variability, but increase the intra-task variability. This reduces the
sensitivity of the MWU test, so bump up the significance level.

In the future, I think we want to hard-code the significance level at 0.001,
but increase the number of repeats for regressions that are smaller compared
to the noise. We can also do more repeats if some of the tests fail.

Review-Url: https://chromiumcodereview.appspot.com/3007393002