Refactor EnableAllContexts to return all activated contexts
Previously, EnableAllContexts() return the maximum context id.

I believe the API was this way because in the past, the context ids are guaranteed to
be consecutive number, hence just getting the maximum context would be sufficient
to infer all the context ids. This is guarantee is no longer held, so  this CL refactors it
to return all the activated context ids, which make it easier for client to access all.

(Context: this is part of the work required for https://codereview.chromium.org/3017573002/)

BUG=chromium:763119

Review-Url: https://chromiumcodereview.appspot.com/3017583002