trace-viewer: Let the gzip_importer use Pako directly instead of JSZip
Trace-viewer's version of jszip is stuck at an older version which can't
be easily upgraded due to newer versions only supporting async operations.

In addition, the newer versions of JSZip have deprecated several utility
functions used by the gzip_importer (to do string<->array conversions and
type checking); these have been replaced by manual versions to more easily
be able to upgrade to a new JSZip version in the future. JSZip already uses
Pako internally so this shouldn't affect performance or functionality.

The main motivation behind this is to be able to continue using Catapult for
embedders which are pegged at a newer version of JSZip.

After this, the only remaining use of JSZip is the zip_importer.html.

Change-Id: I988031d6743c93a964aa9a45cf2a60c3b1c2e802
Reviewed-on: https://chromium-review.googlesource.com/699022
Reviewed-by: Ben Hayden <benjhayden@chromium.org>
Reviewed-by: David Tu <dtu@chromium.org>
Commit-Queue: David Tu <dtu@chromium.org>