Fix add_reserved_diagnostics
This script is currently hitting ValueError('Cannot add same Histogram twice').
This suggests that two Histograms have the same guid.
It turns out that merge_histograms_cmdline.html will only clone() Histograms
(thereby generating a new guid) if there are multiple Histograms that need to be
merged together. So, in add_reserved_diagnostics, `dicts_across_repeats` and
`dicts_across_stories` each contain a single Histogram with the same guid,
causing the ValueError.

This CL changes add_reserved_diagnostics to detect this case and drop one of the
duplicates.

While debugging this issue, I also noticed that sometimes shared diagnostics
would be serialized as a guid reference in the Histogram's diagnostics, but fail
to be serialized in full by the HistogramSet because there is only 1 Histogram
that refers to them (count is not greater than 1 in
HistogramSet.deduplicateDiagnostics()).

This CL changes deduplicateDiagnostics to add a diagnostic to the HistogramSet
if either count > 1 or the diagnostic already has a guid.
In our case, the diagnostics already have guids because they were initially
added as shared diagnostics, and their guids were not reset because their
referencing Histogram was not cloned in merge_histograms_cmdline.html.

See the bug for a demo of how this script now handles the offending
perftest-output.json.

https://github.com/catapult-project/catapult/issues/4001
Bug: catapult:#4001

TBR=simonhatch@chromium.org

Change-Id: Iee283e7c6e3a2e5eda531f43530dbaacffa7db57
Reviewed-on: https://chromium-review.googlesource.com/741444
Reviewed-by: Ben Hayden <benjhayden@chromium.org>
Commit-Queue: Ben Hayden <benjhayden@chromium.org>