[results.html] Move the Export menu into a tr-ui-b-dropdown.
The Export menu is currently a hacky re-implementation of a dropdown.
This CL refactors the Export menu to re-use tr-ui-b-dropdown, allowing
deletion of some code.

The only visible change is that the buttons in the menu are now in a
grid and styled like Material Flat Buttons:
https://material.io/guidelines/components/buttons.html#buttons-flat-buttons

Demo: http://www/~benjhayden/742705.html

Change-Id: I96febeec0b3d6926fdd00e09f3c6cda555c9246e
Reviewed-on: https://chromium-review.googlesource.com/742705
Reviewed-by: Simon Hatch <simonhatch@chromium.org>
Commit-Queue: Ben Hayden <benjhayden@chromium.org>