Rename Time To Interactive metrics
To simplify developer messaging, we renamed Time to Consistently
Interactive to just Time to Interactive, and Time to First Interactive
to Time to First CPU Idle. This CL renames variables and functions
appropriately in the time_to_interactive module before we wire them up
to loading_metric.

There are lots of docs lying around that still mention the term "First
Interactive" and "Consistently Interactive", and it will be a while
until we can rename everything appropriately. For the time being, I left
comments in the jsdoc about the previous names of these metrics.

Bug: catapult:#3635
Change-Id: I408aa1cf201a9e4f264697a36a94f9e161dece8c
Reviewed-on: https://chromium-review.googlesource.com/774987
Reviewed-by: Timothy Dresser <tdresser@chromium.org>
Reviewed-by: Ben Hayden <benjhayden@chromium.org>
Commit-Queue: Deepanjan Roy <dproy@chromium.org>