Add 'CompositorThread' to telemetry GPU thread names.
When running with --enable-viz code is shuffled around between
processes. Some trace events for the smoothness benchmark are output in
the GPU process on the 'CompositorThread' with --enable-viz. Other GPU
threads don't output trace events. Telemetry doesn't look for
'CompositorThread' and misses all trace events from the GPU process as a
result.

Update TraceEventTimelineImporter to look for 'CompositorThread' when
it's trying to identify the GPU process.

Bug: chromium:778749
Test: Verified telemetry works with --enable-viz locally.
Change-Id: Iea2bcc1cdc412c1ea39e0d3c0f008977c3725563
Reviewed-on: https://chromium-review.googlesource.com/806717
Reviewed-by: Ned Nguyen <nednguyen@google.com>
Reviewed-by: Timothy Dresser <tdresser@chromium.org>
Commit-Queue: kylechar <kylechar@chromium.org>