Make pinch actions take speed in viewport coords
I was a little over-zealous in
https://crrev.com/37ab2bfdeb6d954a88088cc163675b5de67afc5b. I made the
speed parameter in PinchAction be in CSS pixels. This makes sense for
scrolls and other gestures but for pinch it means that the speed of the
zoom depends on how far the page is currently zoomed in. This is never
what you want for pinch-zoom.

I've updated pinch.js to keep the speed invariant across page scale
factors. I've also added a test for this and other pinch zoom aspects
since we currently have poor test coverage.

Bug: chromium:610021
Change-Id: I7008cb6c90aa4b9a610abd2c322c88281aa97f87
Reviewed-on: https://chromium-review.googlesource.com/783823
Reviewed-by: Ned Nguyen <nednguyen@google.com>
Commit-Queue: David Bokan <bokan@chromium.org>