Catapult: Manifest parsing now works with xml namespaces stripped
XML namespaces are not used at runtime and thus are not required in the final
apk. Currently the manifest parser bugs out if there is no namespace on the
first line in the manfiest file aapt dump output. This CL fixes this and allows
the parser to work with or without namespaces. It also adds some comments on
the manifest parsing function.

Bug: catapult:#4133
Change-Id: I5dac2d91c465ead35d5ded73cab6e8a7dde7bc53
Reviewed-on: https://chromium-review.googlesource.com/837940
Reviewed-by: John Budorick <jbudorick@chromium.org>
Commit-Queue: Mohamed Heikal <mheikal@chromium.org>