If a sync request to DevTools times out, try to crash the renderer process.
This uses a new DevTools API being introduced here:
  https://chromium-review.googlesource.com/858379

It's not clear whether this is the cause of the random WebGL conformance
test hangs. However, if the main thread of the renderer process is hung,
this change will provide symbolized stack traces for all threads, by virtue
of producing a minidump for the hung renderer process.

BUG=chromium:797368

Change-Id: I079a1d3343002b7f2dc37419aeb30f07e9f7a1a3
Reviewed-on: https://chromium-review.googlesource.com/857959
Commit-Queue: Kenneth Russell <kbr@chromium.org>
Reviewed-by: Ned Nguyen <nednguyen@google.com>