[Dashboard] Ignore reference build directory when deploying
The reference build is too big to deploy and is not required by the
dashboard. It is, however, in the py_utils/py_utils directory, which we
need, so our best course of action is to use app.yaml's skip_files option,
as suggested by the deployment script. Closes #4194.

Bug: catapult:#4194
Test: dashboard/bin/deploy
Change-Id: I03c7b1456882a00ff452790f50e2112838297686
Reviewed-on: https://chromium-review.googlesource.com/884235
Commit-Queue: Ethan Kuefner <eakuefner@chromium.org>
Reviewed-by: Simon Hatch <simonhatch@chromium.org>