[results.html] Auto Show All when empty search results.
Currently, the table is empty when the user searches for a non-top-level
histogram name when Show All is unchecked.
That behavior can lead users to incorrectly assume that the metric failed to
record the histogram.

This CL ensures that users can find the desired histogram if it exists by
detecting when the search results would be empty because Show All is unchecked,
and automatically checks it.

Demo of fix: http://www/~benjhayden/801556.html

Bug: chromium:801556
Change-Id: I02060cbd5c013a05975bd1a027ff2b91c0a7362a
Reviewed-on: https://chromium-review.googlesource.com/893904
Reviewed-by: Ethan Kuefner <eakuefner@chromium.org>
Commit-Queue: Ben Hayden <benjhayden@chromium.org>