[Telemetry] Add ShouldAddValue hook to replace ValueCanBeAddedPredicate
This CL adds a hook ShouldAddValue that is intended to replace
ValueCanBeAddedPredicate. It takes value names instead of value instances
so that it can also be used to filter histograms.

This is the beginning of a two-sided patch, so a subsequent patch will update
overrides in Chromium of ValueCanBeAddedPredicate to equivalent ShouldAddValue
implementations. Once that lands, final Catapult CLs will remove references to
ValueCanBeAddedPredicate and begin using ShouldAddValue to filter histograms.

Bug: catapult:#4215
Change-Id: I9efd3927868192f659700003f3c41dd06954396c
Reviewed-on: https://chromium-review.googlesource.com/896688
Commit-Queue: Ethan Kuefner <eakuefner@chromium.org>
Reviewed-by: Ned Nguyen <nednguyen@google.com>