[Telemetry] Fix bug with filtering histograms via ShouldAddValue
Currently, we apply ShouldAddValue to histogram names directly, which is
sometimes not quite right. For example, several TBMv2 benchmarks filter using
a regex which excludes names ending in "count", but this leads to "dump_count"
being filtered out (when really you'd want "dump_count_count" to be filtered
instead).

To work around this, and allow users to still filter histograms that they
don't want to be present in the final results, we instead check whether
ShouldAddValue filters _all_ statistics, and if it does, then we do not add
the histogram. This allows filtration based on 'in' checks like smoothness and
some other legacy benchmarks currently do, to be performed for TBMv2 benchmarks
as well, while avoiding the situation where we simply filter out "dump_count"
because it ends with count.

Bug: catapult:#4268
Change-Id: I0fdc02661b3fd64add28338a4e36659bff72c333
Reviewed-on: https://chromium-review.googlesource.com/923098
Commit-Queue: Ethan Kuefner <eakuefner@chromium.org>
Reviewed-by: Ned Nguyen <nednguyen@google.com>