Pinpoint - Have results2 generated on the backend.
Previously, we tried to download all the histograms for a job and then
render out the vulcanized view. This would fail on large datasets with
memory issues. Additionally, it always tried to render from scratch.

This restructures the endpoint to generate it in a task on the task
queue (which has a much higher timeout limit). We don't try to download
all the histograms at once, instead we stream them in one by one as
we render out the viewer, which is streamed out to cloud storage.

Bug: catapult:#4168
Change-Id: Ide282d20abd8e050f35e9d83632833ecee651378
Reviewed-on: https://chromium-review.googlesource.com/935146
Reviewed-by: David Tu <dtu@chromium.org>
Commit-Queue: Simon Hatch <simonhatch@chromium.org>