Ignore i-frame navigationStart events when computing LoadExpectation
We seem to have navigationStart events corresponding to i-frames that
are reported as navigationStarts for main frame with a blank url ('').
These events have to be ignored, so that we don't have overlapping
loading expectations.

Also use load expectations to compute TTI instead of using loading
metric. This would be cleaner and remove dependency on loading metric.

Bug:chromium:796109
Change-Id: I7edf5aa0b9f32082539df7114aeaf8014592852d
Reviewed-on: https://chromium-review.googlesource.com/937701
Commit-Queue: Mythri Alle <mythria@chromium.org>
Reviewed-by: Ben Hayden <benjhayden@chromium.org>
Reviewed-by: Deepanjan Roy <dproy@chromium.org>