[Dashboard] Add IS_SUMMARY diagnostic
This CL reserves a new IS_SUMMARY diagnostic and checks for it when computing
test paths in /add_histograms. It sets this diagnostic in
add_reserved_diagnostics for histograms merged across stories that only refer
to one story, so that for story sets that only have one story, we still produce
both a summary test and a per-page test, on the dashboard side.

Bug: catapult:#4327
Change-Id: I3ad3f26b87644f463fb795c7f071575d70b20c1c
Reviewed-on: https://chromium-review.googlesource.com/959349
Commit-Queue: Ethan Kuefner <eakuefner@chromium.org>
Reviewed-by: Simon Hatch <simonhatch@chromium.org>