Fix service starting on N and below
Starting services on N and below is done via "am startservice", while
on O and above, it's done via "am start-service". When service starting
was initially added, it was only tested on O, so the current
implementation was failing on N and below due to am not knowing what
"start-service" was.

Bug: chromium:828190
Change-Id: I3444563d66beb8ec0accbe45878be98035e43a4f
Reviewed-on: https://chromium-review.googlesource.com/998516
Reviewed-by: Benjamin Pastene <bpastene@chromium.org>
Commit-Queue: Brian Sheedy <bsheedy@chromium.org>