Dashboard - Refactor /migrate_test_names
Migration is pretty slow, on the order of days for tests with a lot
of data, hoping to speed it up a bit.

Previously, the endpoint worked by migrating everything in a single
task. This changes the functionality to spawn a separate task for each
test_path.

Bug: catapult:#4149
Change-Id: Ia466f28c18c5292ba3de72a9d40e1dbd5892e561
Reviewed-on: https://chromium-review.googlesource.com/1012726
Reviewed-by: Annie Sullivan <sullivan@chromium.org>
Reviewed-by: Ethan Kuefner <eakuefner@chromium.org>
Commit-Queue: Simon Hatch <simonhatch@chromium.org>