Dashboard - Automate TimeToCulprit calculation.
This metric is part of the Chrome Project Health Metrics. It's the time
between a commit landing and a culprit being identified (by Pinpoint).
Eventually, we'll break this down further, time from landing to alert,
from alert to job creation, and finally to culprit. For now, this just
covers the entire range.

Bug: chromium:860334
Change-Id: I7ed6ad0d2b687180fd11be242c48498097fd773a
Reviewed-on: https://chromium-review.googlesource.com/1126213
Reviewed-by: Ethan Kuefner <eakuefner@chromium.org>
Commit-Queue: Simon Hatch <simonhatch@chromium.org>