[ftrace importer] Fix recursive binder reply arguments
Currently when there is a recursive binder transaction, the reply
slice for the orignal transaction has erroneous arguments. Instead
of having arguments indicating a reply back to the original slice,
the arguments are identical to those of the first recursive
transaction.

eg. if you have threads A, B, C, and have transactions:
A->B->C->B->A, then the final "B->A" reply slice will have the same
arguments as the "B->C" transaction slice, which is incorrect.

This change modifies the behavior to set the arguments of the reply
slice so that it indicates that it is performing a transaction back
to the original thread.

Change-Id: I433800c02a58ce3f08d75897436e4622bd80be50
Reviewed-on: https://chromium-review.googlesource.com/1130951
Reviewed-by: Ethan Kuefner <eakuefner@chromium.org>
Commit-Queue: Ethan Kuefner <eakuefner@chromium.org>