Integration
Here I integrated the plot feature within the drop-down menu: after
the user chose a specific metric + story + diagnostic, a plot will be
displayed for all the values of different diagnostics for these
specific elements; in other word, the plot gives a comparison
of different values for each type of X diagnostic.

Currently the data is displayed in a table; this is provided with
some additional features: choose how to sort the inputs from table
and search for a specific name of metric/story/diagnostic and
then have an entry of all results.

BUG:chromium:866423
Change-Id: I25d9d0fa0530d15a006d6183b71a873899b0c09a
Reviewed-on: https://chromium-review.googlesource.com/1164959
Commit-Queue: Florina Năstăsoiu <nastasoiuf@google.com>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>