[dashboard] Update docs to mention old issues filed in the Github tracker
The docs should mention the Github issue tracker so that anyone
looking for existing filed issues can know to look into both the
Chromium and Github issue trackers.

Also mention the component name in the text and provide a link for
searching for issues.

Change-Id: Ic8eb0f70df1364bdf5cd3a769b43bd13c27b83fb
Reviewed-on: https://chromium-review.googlesource.com/1180355
Reviewed-by: David Tu <dtu@chromium.org>
Commit-Queue: Mark Seaborn <mseaborn@chromium.org>