Improve legends for charts.
This CL improves the styling of legends
by using black text to improve readablity and adding
colored graphics to denote labels. It also increases the size
of the chart and the size of the legend area to make use of available
whitespace in results.vue, reducing the chance of text going off chart.

Bug: chromium:866423
Change-Id: I2ab91213f3aef0c9ac242b8a68911db28828a51a
Reviewed-on: https://chromium-review.googlesource.com/1198534
Commit-Queue: Anthony Alridge <anthonyalridge@google.com>
Reviewed-by: Mythri Alle <mythria@chromium.org>