devil: Limit CPU/GPU frequency for Nexus 5X
Magic parameters arrived for N5X benchmarking mode. Selection of them is
not very scientific an probably not optimal - just lowering the max
allowed frequency by one stop (like the N5). The N5X has big cores,
which can be higher frequency, maxing them to the same frequency as the
small cores, for extra stability. The choice for big cores may be not
ideal for representativeness, also they'd still be a little faster than
the small ones.

On N5X the cores 0-3 are little, 4-5 are big. The easiest way to confirm
this is by looking at:
    /sys/devices/system/cpu/cpuX/cpufreq/scaling_available_frequencies

Similarly for GPU - one stop higher than the lowest. The list of
available GPU frequencies is available here:
    /sys/class/kgsl/kgsl-3d0/gpu_available_frequencies

To test how this choice works I ran experimental.startup.mobile with a
couple of customizations:

1. Doing WaitForBatteryTemperature() only once before all other
   benchmarks. This is to confirm that the benchmark can run well
   without cooling delays (those are slow). To confirm that startups by
   the end of the storyrun are not obviously slower than the ones in the
   beginning. In ideal worlf we could more reliably detect cases when
   the CPU gets throttled, but this may introduce overhead, and may
   itself heat the device up - good to be lazy sometimes.

2. Starting only when below 32C (~90F) - based on previous knowledge,
   this battery temperature 'usually' provided faster startups. The
   current 35C is a little too hot for benchmarking, even though may be
   good for preventing battery explosions. Again, not scientific :)

3. 20 runs instead of 10 for more confidence

The traces are here: http://crbug.com/383566#c42

Bug: chromium:760498
Bug: chromium:383566

Change-Id: I9cf2e197d8d93281a7a0422269765d35dade9240
Reviewed-on: https://chromium-review.googlesource.com/1196555
Commit-Queue: Egor Pasko <pasko@chromium.org>
Reviewed-by: Ned Nguyen <nednguyen@google.com>