Add ResultChannel{Sender,Receiver} for v2spa
This pair of classes is used to stream timeseries and report data from the
service worker to the page. A following CL will add TimeseriesCacheRequest and
ReportCacheRequest, which read data from IndexedDB and one or more API fetches,
and forwards data to the page as it becomes available.

Bug: catapult:#4461
Change-Id: If413a54a63167c146e412741d8232153f1197128
Reviewed-on: https://chromium-review.googlesource.com/c/1264521
Commit-Queue: Ben Hayden <benjhayden@chromium.org>
Reviewed-by: Simon Hatch <simonhatch@chromium.org>