Enable orderfile memory optimization in benchmarks
**PERF SHERIFFS**: We do not expect the behavior to affect startup
                   performance. It should affect memory usage from
                   native code.

The way to force the new behavior (that we intend to launch) is to
provide the flag to Chrome: --orderfile-memory-optimization.

This behavior will be Finch-ed (as described in [1]), but we cannot use
the 'fieldtrial_testing_config.json' for testing because the feature is
set up as a synthetic field trial.

This change should land and rolled into Chromium _before_ [1] lands to
avoid coin flipping when benchmarking.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/1270936

Bug: chromium:758566
Change-Id: I5963b087b0226e15befbd68335325147cd3efeba
Reviewed-on: https://chromium-review.googlesource.com/c/1276246
Reviewed-by: Juan Antonio Navarro Pérez <perezju@chromium.org>
Commit-Queue: Egor Pasko <pasko@chromium.org>