[telemetry] Treat most exceptions as "possibly handleable"
Telemetry classifies errors as "handleable" and "unhandleable" errors.
Handleable errors, defined by a short explicit list, are displayed and
reported as an individual story failure. Unhandleable errors, any
error not in the list, are treated as fatal errors and stop the entire
benchmark execution.

This is often too drastic, as many errors raised by Telemetry clients
are in fact "handleable" and should not stop the execution of the
entire benchmark due to an error in one of its stories.

In this CL we define instead a small set of explicitly "unhandleable"
errors (KeyboardInterrupt, SystemExit, etc) which should definitely
halt the benchmark execution, while most other Exception's are
treated as "possibly handleable", allowing other stories to continue
running but restarting the shared state to be safe.

Bug: chromium:900909
Change-Id: I5f23d5c33319b04f7631f0f3e77fe2a2bb62c131
Reviewed-on: https://chromium-review.googlesource.com/c/1326943
Reviewed-by: Caleb Rouleau <crouleau@chromium.org>
Reviewed-by: Ned Nguyen <nednguyen@google.com>
Commit-Queue: Juan Antonio Navarro Pérez <perezju@chromium.org>