Add Range as es6 module for v2spa.
This is required for TimeseriesCacheRequest:
https://chromium-review.googlesource.com/c/catapult/+/1290176

This is mostly copied from tracing because es6 modules like the service
worker can't import from html files. Eventually, the Histogram system
and its dependence on Range will be rewritten to web assembly in a way
that the service worker can import.
https://github.com/catapult-project/catapult/blob/master/tracing/tracing/base/math/range.html

Bug: catapult:#4461
Change-Id: I3cf5a9281bc6901bb4532137f6ad5d1f21686023
Reviewed-on: https://chromium-review.googlesource.com/c/1298593
Reviewed-by: Sean McCullough <seanmccullough@chromium.org>
Commit-Queue: Ben Hayden <benjhayden@chromium.org>