[soundwave] Use explicit dtype's for DataFrame columns
Otherwise letting pandas guess the dtypes is brittle when, e.g., an API
returns no rows of data.

For this we adapt the existing _EmptyDataFrame used to create the
database table schemas and adapt it to be able to also receive rows
of data.

Bug: chromium:904876
Change-Id: I69ef496e2b1ae33a516a219ff27dbec1181528d0
Reviewed-on: https://chromium-review.googlesource.com/c/1341929
Commit-Queue: Juan Antonio Navarro Pérez <perezju@chromium.org>
Reviewed-by: Ben Hayden <benjhayden@chromium.org>