Remove precision check from the EQT metric calculation
The current 0.1ms precision check for floating point errors
starts failing once the trace events are 30s into the trace
(or further). As the times of the events are adjusted to
split the difference, this check is not needed (and
would always fail with big enough values).

R=ulan@chromium.org,benjhayden@chromium.org

Change-Id: I1c26ffb9bd442488d2bdb44d69dc06bd72fac03b
Reviewed-on: https://chromium-review.googlesource.com/c/1364251
Commit-Queue: oysteine <oysteine@chromium.org>
Reviewed-by: Ulan Degenbaev <ulan@chromium.org>
Reviewed-by: Ben Hayden <benjhayden@chromium.org>
Auto-Submit: oysteine <oysteine@chromium.org>