[Telemetry] Control default action timeout in a single constant.
It seems like 60 seconds, which is what everyone seems to be using,
is a bit longer than it should be on modern computers. A high
timeout means that we are more likely to hit suite-level timeouts
and we are more likely to waste time waiting for a broken story
rather than moving on to stories that work correctly. In a
followup CL, I will see if I can reduce this timeout without breaking
a lot of tests. Any tests that break can be edited to not use the
default.

Bug: chromium:921662
Change-Id: Ib9ba75da6b2041200d07467306fa583aeb6e4613
Reviewed-on: https://chromium-review.googlesource.com/c/1413774
Commit-Queue: Caleb Rouleau <crouleau@chromium.org>
Reviewed-by: John Chen <johnchen@chromium.org>