Add APIs to cros interface and cros platform backend class.
Add a cros interface API to start a command on the CrOS device and return the
process handle. This API will be used by the browser_interval_profiling_controller
to start the CPU profiler during story runs.

Add cros platform backend APIs to expose the cros interface APIs to start a
command on the CrOS device, get a file from the CrOS device, push contents to a
file on the CrOS device.

Fix the doc string of the GetFile API.

Test: tested with:
        `./run_tests --browser=cros-chrome --remote=$CROS_REMOTE CrOSInterfaceTest.testStartCmdOnDevice`
        `./run_tests --browser=cros-chrome --remote=localhost --remote-ssh-port=9222 CrOSInterfaceTest.testStartCmdOnDevice`
        `/usr/local/telemetry/src/third_party/catapult/telemetry/bin/run_tests CrOSInterfaceTest.testStartCmdOnDevice`

Change-Id: I94fe4e625cbb57049382c845f26a401bcea0fa6e
Reviewed-on: https://chromium-review.googlesource.com/c/1436219
Commit-Queue: Lakshman Annadorai <lakshmana@chromium.org>
Reviewed-by: Caleb Rouleau <crouleau@chromium.org>
Reviewed-by: Achuith Bhandarkar <achuith@chromium.org>
Reviewed-by: Juan Antonio Navarro Pérez <perezju@chromium.org>