Add a drawer to chromeperf-app
This CL replaces the button bars with a drawer.
This drawer has two states: compressed and expanded.
In the compressed state, only icons are shown.
In the expanded state, labels appear next to the icons.
The drawer is expanded while the mouse hovers over it.
There is no menu button. The drawer can never be complete hidden.

Bug: chromium:918193
Change-Id: Ie78f69110727841c66b530587448db7a50248ed0
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/1497376
Commit-Queue: Ben Hayden <benjhayden@chromium.org>
Reviewed-by: Tiffany Zhang <zhangtiff@chromium.org>