Pinpoint - Fix pinpoint spamming bugs on frozen jobs.
When a job freezes, there's a cron job that attempts to find and retry
or fail these jobs. After a # of retries, the job is failed and not
retried, but this posts to the bug so after the retry limit don't
touch the job again.

Bug: chromium:943054
Change-Id: I2c3a44bd1df68eac66c7dc443505d2c1d3b21f21
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/1529131
Reviewed-by: Ben Hayden <benjhayden@chromium.org>
Commit-Queue: Simon Hatch <simonhatch@chromium.org>