Add sparkline-compound for spa
This component contains a tab strip of sparklines.

Clicking on a sparkline opens a new chart-section containing the selected
timeseries.

Building the sparklines is currently specific to chart-section with its
timeseries-descriptor, but could be generalized to also support the
chart-compound in alerts-section.
Sparklines can be built pro-actively, and are rendered lazily when the user
opens a tab.

X- and Y-axis cursors track the user's cursor over a main chart.
Sparklines' revision ranges update when the user brushes a minimap.

Bug: chromium:918193
Change-Id: I90aecac8279f9e9d04fc7a5578ea620ce149aead
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/1526550
Reviewed-by: Simon Hatch <simonhatch@chromium.org>
Commit-Queue: Ben Hayden <benjhayden@chromium.org>