rendering: Remove breakdowns for thread-time metrics.
The breakdowns aren't useful for most thread-times metrics. For
example, thread-time for display-compositor, or renderer-compositor
only contain data for the single thread. So for these cases, we do
not need the breakdown metrics. Remove such breakdowns, and leave
the breakdowns only for the total-thread-time and total-fast-path
metrics, since multiple threads contribute to these metrics.

BUG=chromium:948168

Change-Id: Ie92bebfa13ce661b610979e1f2a8e3d4df2acc26
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/1547951
Reviewed-by: Caleb Rouleau <crouleau@chromium.org>
Commit-Queue: Sadrul Chowdhury <sadrul@chromium.org>