[Telemetry] Clean up local server process handle to remove red herring.
ps_util.ListAllSubprocesses sometimes finds this process still running on
Windows, and that is a red herring since the process will die as soon as
the handle is closed. To fix this, poll() to close the process handle.

Note that this is flaky because if garbage collection happens then that
would close the handle for us, so that is why this usually works fine.

Bug: chromium:953365
Change-Id: I3d5a3edb8afaeae55e3406a8046b7b07f4c9f5d3
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/1573139
Auto-Submit: Caleb Rouleau <crouleau@chromium.org>
Reviewed-by: John Chen <johnchen@chromium.org>
Commit-Queue: John Chen <johnchen@chromium.org>