SheriffRedux: Sheriff Config Service
This CL implements the beginnings of a service to be hosted in the
chromeperf dashboard project, which has the following properties:

- Uses the `python37` runtime environment for AppEngine
- Uses Flask as the web application framework
- Depends on gunicorn for WSGI serving
- Uses pip dependencies in the requirements.txt file

This change allows us to isolate the service which performs the
validation and storage of the configuration we're going to be receiving
through API calls from the luci config service. The isolation allows us
to implement the validation logic in a single place, and let the
dashboard use the configuration received through the service.

Isolating the service from the rest of the dashboard also allows us to
independently deploy the service from the rest of the dashboard, and let
the dashboard use the configuration through a well defined API. It also
gives us the chance to use an upgraded Python environment, manage the
dependencies explicitly, and have a smaller testing surface for the
service.

R=simonhatch@chromium.org

Bug: chromium:953181
Change-Id: I2eafe580994738f2dee692c39b5a02ea68ac5372
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/1578618
Auto-Submit: Dean Berris <dberris@chromium.org>
Reviewed-by: Simon Hatch <simonhatch@chromium.org>
Commit-Queue: Dean Berris <dberris@chromium.org>