[Telemetry] Set Telemetry info attributes in constructor
Deprecate the old way in which clients needed to write on attributes;
as that was harder to code search.

Also, since the PageTestResults object already knows about the
benchmark metadata, there is no need for clients to supply this
information twice.

TBR=crouleau@chromium.org
Bug: chromium:973837
Change-Id: I6c8587150523d1eb546d8d24a47bf4aa146265cf
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/1662283
Reviewed-by: Juan Antonio Navarro Pérez <perezju@chromium.org>
Commit-Queue: Juan Antonio Navarro Pérez <perezju@chromium.org>