[Devil] Double APK install timeout to 4 minutes
Installing app bundles takes longer than APKs. It appears that we
were close to the 2 minute timeout on the android-pixel2-perf-fyi
bot on the successful build, and now the bots are failing due to
this step timing out. Increase the timeout to see if this fixes
the issue.

BUG=chromium:934251

Change-Id: I7455c44a417b238b91ba4a9ccb6ea4388ed7a230
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/1677205
Auto-Submit: Ross McIlroy <rmcilroy@chromium.org>
Commit-Queue: Juan Antonio Navarro Pérez <perezju@chromium.org>
Reviewed-by: Juan Antonio Navarro Pérez <perezju@chromium.org>