[LCP] Support LargestContentfulPaint for Telemetry
After landing crrev.com/c/1671027, Telemetry will be able to get
the LargestContentfulPaint from the browser events:
1. "NavStartToLargestContentfulPaint::Candidate::AllFrames::UKM"
2. "NavStartToLargestContentfulPaint::Invalidate::AllFrames::UKM"

We do not calculate LCP from the existing LargestImagePaint and
LargestTextPaint because they are for mainframe. The UKM-LCP is for
all frames so it's closer to what UKM reports.

Bug: chromium:966453, chromium:973836

Change-Id: I7e0d86d88e09161c5cc20009258fc627d21e56d9
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/1672444
Reviewed-by: Deepanjan Roy <dproy@chromium.org>
Reviewed-by: Ben Hayden <benjhayden@chromium.org>
Commit-Queue: Liquan (Max) Gu <maxlg@chromium.org>