PinpointScheduler: Link to jobs in queue
This change exposes the list of jobs associated with a particular queue,
to allow users to navigate to individual jobs, along with the status
associated with those as well. We also update the style associated with
the SVG of the graph to be consistent with the other page.

TODO: In the future we shoudl consolidate the histogram visualisation
across the various "stats" pages across Pinpoint.

R=abennetts@google.com, simonhatch@chromium.org

Bug: chromium:977488
Change-Id: I6993dfa1dd6ebb3f728ef1ae7667e6335e6d26a4
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/1712924
Auto-Submit: Dean Berris <dberris@chromium.org>
Reviewed-by: Simon Hatch <simonhatch@chromium.org>
Commit-Queue: Dean Berris <dberris@chromium.org>