PinpointScheduler: Post cancellation comments
This change makes Pinpoint update bugs for which Pinpoint runs have been
cancelled, and only to those that have a bug ID associated with them.
Ideally we should also send emails, but if we're doing that we should
ensure that those emails are getting more information than just what
we're posting in the bugs as comments.

We can refine the information we send as email updates independent of
bugs being associated with a job in later changes.

R=abennetts@google.com, simonhatch@chromium.org

Bug: chromium:977488
Change-Id: Ia79da5fa3afe3c38d21ffd1aa8dc5aea801be0ca
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/1712927
Reviewed-by: Andrew Bennetts <abennetts@google.com>
Commit-Queue: Dean Berris <dberris@chromium.org>