Pinpoint: Update stats to account cancelled jobs
This change allows us to count cancelled jobs in the stats page, and
more accurately count the failed jobs to not include the 'Queued' jobs.
We update the javascript code to more closely reflect our expectations
when performing the analysis of the data we're getting from the service.

R=abennetts@google.com, simonhatch@chromium.org

Bug: chromium:977488
Change-Id: I25f6b07642431b446347c497692a87aaad5e0cde
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/1724718
Reviewed-by: Simon Hatch <simonhatch@chromium.org>
Reviewed-by: Andrew Bennetts <abennetts@google.com>
Commit-Queue: Dean Berris <dberris@chromium.org>