[Telemetry] Do not depend on sys.argv to sniff benchmark name
The old code in Run.AddCommandLineArgs used to scan sys.argv to
find possible benchmark names, to then given them a chance to
define extra command line arguments to parse.

This made it hard to write tests for the command_line.parser, as
the `args` it was processing may not correspond to the `sys.argv` of,
say, the test runner script.

Instead, in this CL, we make the dependency of Run.AddCommandLineArgs
to the args being parsed explicit by passing them as an extra argument
to the class method.

We also remove both List and Run commands out of the
command_line.OptparseCommand hierarchy, and make them subclass instead
from object directly. The dependency from command_line.OptparseCommand
did not provide any benefits (virtually no code being reused), and
only complicated things as the signature from the method would
change from that of the parent. (Which, coincidentally, the parent
class already does! [1], so keeping them in the hierarchy would only
complicate things even more.)

Also: Eventually we want to get rid of command_line.OptparseCommand
entirely, so this is as step in the right direction.

[1]: https://cs.chromium.org/chromium/src/third_party/catapult/telemetry/telemetry/internal/util/command_line.py?rcl=bb774e5ede14bc59f24ea9a8d133e59f9c8a3032&l=69

Bug: chromium:985712
Change-Id: I1f862ec039a9c696c5d4cf0923ae5dc56395440c
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/1736651
Commit-Queue: Juan Antonio Navarro Pérez <perezju@chromium.org>
Reviewed-by: Caleb Rouleau <crouleau@chromium.org>