Reland "[Telemetry] Remove StoryExpectations class which was used for the old expectations format"
This is a reland of d69a20cce1f35c0208b5de85fe011341b3d6e530

It was reverted originally so that we could quickly revert the
expectations.config file format changes if need be. (Since we
discovered that there was a bug in it, so the worry was that there
would be more bugs and we would want to revert quickly).

It is now being relanded since the expectations.config changes
seem to be sticking now and we would like to keep building on
top of this.

Original change's description:
> [Telemetry] Remove StoryExpectations class which was used for the old expectations format
>
> Since we are removing this class, we can also remove useless arguments to
> typ_expectations.StoryExpectation's IsDisabled and IsStoryDisabled methods.
>
> Bug: chromium:973936
> Change-Id: I31ef59d86624d0341dc45140e2568a3ed261cdb9
> Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/1745353
> Commit-Queue: Rakib Hasan <rmhasan@google.com>
> Reviewed-by: Juan Antonio Navarro Pérez <perezju@chromium.org>

Bug: chromium:973936
Change-Id: I6ea3c059f1c5315733b123c0ef24b1777d4d8cae
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/1772822
Commit-Queue: Caleb Rouleau <crouleau@chromium.org>
Reviewed-by: Rakib Hasan <rmhasan@google.com>
Reviewed-by: Juan Antonio Navarro Pérez <perezju@chromium.org>
Reviewed-by: Caleb Rouleau <crouleau@chromium.org>