PinpointScheduler: Backwards compatibility for find_isolate
Instances of the find_isolate and run_test objects could have been
stored to not have the requisite additional fields, so we should test
first whether the newly added fields are in the object before accessing
them. This is necessary because Pinpoint pickles these objects and
stores those in Datastore, and therefore could be seeing a previous
version of the object stored and reconstituted in the live service.

We learned this on a production deployment, where ongoing jobs started
failing due to an unrecognised field in the FindIsolate type.

R=abennetts@google.com, simonhatch@chromium.org

Bug: chromium:988912
Change-Id: I854f2c72ae3eefa096d7962245551f53d3a4ce62
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/1758145
Commit-Queue: Dean Berris <dberris@chromium.org>
Auto-Submit: Dean Berris <dberris@chromium.org>
Reviewed-by: Simon Hatch <simonhatch@chromium.org>