Fix tracing not started problem after Chrome OS is mounted to remove write-protection
As landed in crrev.com/c/1778852, Chrome OS DUT needs to remount the root to remove
write protection. This creates a problem to start tracing. Tracing is started before
the remount, and it writes a file to the DUT. After the remounting, the file is lost
so the tracing appears never to start and causes benchmark run fail. This patch moves
the write protection check up to write the tracing configuration file, so it will
remount the root and then write the tracing file to avoid lost.

Bug: chromium:1000760
Test: Tested locally on Chrome OS device.
Change-Id: I5fd00978b3d3fc16b2bb162d1df6cff46b345d86
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/1815343
Reviewed-by: Brian Sheedy <bsheedy@chromium.org>
Reviewed-by: Juan Antonio Navarro Pérez <perezju@chromium.org>
Reviewed-by: Achuith Bhandarkar <achuith@chromium.org>
Commit-Queue: Tiancong Wang <tcwang@google.com>